Integration tests refactoring
