+++ test/GodelTech.Microservices.IntegrationTests/GodelTech.Microservices.IntegrationTests.csproj
@@ -11,7 +11,7 @@
     <PackageReference Include="xunit" Version="2.4.0" />
     <PackageReference Include="xunit.runner.visualstudio" Version="2.4.0" />
     <PackageReference Include="coverlet.collector" Version="1.2.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="3.1.0"/>
+    <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="3.1.0" />
     <PackageReference Include="FluentAssertions" Version="5.10.3" />
 
   </ItemGroup>
